docker/back/requirements.txt
boto3==1.16.35
botocore==1.19.35
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
diff-match-patch==20200713
dill==0.3.3
Django==2.2.17
django-environ==0.4.5
django-import-export==2.4.0
django-phonenumber-field==5.0.0
django-storages==1.10.1
et-xmlfile==1.0.1
Flask==1.1.1
Flask-Login==0.5.0
gunicorn==19.9.0
idna==2.8
importlib-metadata==3.1.1
itsdangerous==1.1.0
jaconv==0.2.4
Janome==0.4.1
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
json5==0.8.5
klepto==0.2.0
MarkupPy==1.14
MarkupSafe==1.1.1
mecab-python3==0.7
nltk==3.5
numpy==1.17.4
odfpy==1.4.1
opencv-python==4.4.0.46
openpyxl==3.0.5
pandas==0.25.3
phonenumbers==8.12.14
Pillow==7.1.0
pox==0.2.9
psycopg2==2.8.6
pycurl==7.43.0
pygobject==3.20.0
pykakasi @ git+https://github.com/miurahr/pykakasi@7d2179e7ec3fde9af2b5cb653c6475c819142c26
python-apt==1.1.0b1+ubuntu0.16.4.10
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2020.4
PyYAML==5.3.1
qrcode==6.1
regex==2020.11.13
requests==2.22.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.4.1
symspellpy==6.7.0
tablib==3.0.0
tqdm==4.54.1
unattended-upgrades==0.1
urllib3==1.25.11
Werkzeug==1.0.1
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
